:sparkles: Loading画面の実装 fixes #60
