+++ src/app/ui/articles/article/article-content.jsx
@@ -8,7 +8,9 @@ export default function ArticleContent({ article }) {
             <p>{article.body}</p>
             <ul className="tag-list">
               {article.tagList.map((tag) => (
-                <li className="tag-default tag-pill tag-outline">{tag}</li>
+                <li className="tag-default tag-pill tag-outline" key={tag}>
+                  {tag}
+                </li>
               ))}
             </ul>
           </div>
