+++ src/app/ui/articles/feed/feed.jsx
@@ -1,13 +1,24 @@
-import Link from "next/link";
+import { headers } from "next/headers";
+import { fetchArticles } from "@/app/lib/data";
+import { getMaxPage } from "@/app/lib/calculate";
 import ArticleMeta from "@/app/ui/articles/article-meta";
 import Favorite from "@/app/ui/favorite/favorite";
 import FeedContent from "@/app/ui/articles/feed/feed-content";
+import Pagination from "@/app/ui/articles/feed/pagination";
 
-export default function Feed({ articles }) {
+export default async function Feed({ page, limit, offset }) {
+  const pathname = headers().get("x-pathname") || "";
+  const articlesData = await fetchArticles({ page: page });
+  const articles = articlesData.articles;
+  const maxPage = getMaxPage({
+    articlesCount: articlesData.articlesCount,
+    limit: limit,
+  });
+
   return (
     <>
       {articles.map((article) => (
-        <div className="article-preview">
+        <div className="article-preview" key={article.slug}>
           <div className="article-meta">
             <ArticleMeta article={article} />
             <Favorite article={article} />
@@ -16,6 +27,8 @@ export default function Feed({ articles }) {
           <FeedContent article={article} />
         </div>
       ))}
+
+      <Pagination pathname={pathname} page={page} maxPage={maxPage} />
     </>
   );
 }
