+++ src/app/ui/articles/feed/loading.jsx
@@ -1,0 +1,7 @@
+export default function Loading() {
+  return (
+    <>
+      <p className="loader">Loading...</p>
+    </>
+  );
+}
